app_srtm: gpio config input: add explicit irq disable and ignore none
We do not want to stop handling irqs if another process tries to read a
value in between.
Disabling irq will explicitly call us with the new disable value on new
linux versions.

(on older linux versions, we'd just send a notify to linux that would be
ignored, no harm is done)

Change-Id: Ie5c560042bf085c5979b483112308d40594cc2c6