srtm_io_service: remove anti-loop notify logic
When we notify linux we disable the irq anyway, and linux will re-enable
the irq when they are ready to process a new event.

That means there is no risk of spamming the linux side, and more importantly
it was possible to see the re-enable message from linux before the previous
notify had been recycled, making the notification lost and leaving irq
disabled on m33 side without linux knowing about it.

Just always send notifies.

Change-Id: I6ead8f37cab62ee30d6696dfe9b42aaa3f721f74