app_srtm: IO: don't forget about PTA24...
We have PTA[0-24] so 25 pins, add one to APP_IO_PINS_PER_CHIP that
defines the maximum pin value we handle.

Everything was based on this variable so this only needed defining
pinmux to work (thanks to the BUILD_BUG_ON that found it...)
In particular, no linux update is required (max set in dts already
correct)

We normally won't use it as GPIO on A9E (vin ADC input), but better
safe than sorry, and WUU pins need the size to be correct.

Change-Id: Ic0eb9e99484a953073327300b118f23405b78000