rs485: Supports changing the parity, databit and stopbit settings
The specifications are the same as the Linux kernel. 7-bit data
characters without parity, etc. are not supported.

For reference see:
  linux/drivers/tty/serial/fsl_lpuart.c

The format of settings is the same as c_cflag(control mode flags) in
the Linux kernel.

RS485 does not support RTS/CTS flow control. For RS232 you can change
it as follows:
  LPUART_Type *base = LPUART0;
  bool rtscts = tty_rtscts(cflag);
  uint32_t ctrl, modem;

  /* disable transmit and receive */
  ctrl = base->CTRL;
  base->CTRL = ctrl & ~(LPUART_CTRL_TE_MASK | LPUART_CTRL_RE_MASK);
  /* set RTS/CTS flow control */
  modem = base->MODIR & ~(LPUART_MODIR_TXCTSE_MASK | LPUART_MODIR_RXRTSE_MASK);
  modem |= LPUART_MODIR_TXCTSE(rtscts) | LPUART_MODIR_RXRTSE(rtscts);
  base->MODIR = modem;
  /* restore control register */
  base->CTRL = ctrl;

Change-Id: I75a2d932a08630feedbaeb21c9f3949e21a35761