app_pwm: make configuration linux-centric
- always initialize both LPTPM0 and 1
- remove pinmux config

That is all this driver should take with no update on linux side.

If having the TPM core enabled without any channel visibly changes power
consumption then we should consider to initialize the tpm core on first
request in srtm/services/srtm_pwm_service.c, but this is probably good
enough for now.

(Why is this driver implementation different...)

Change-Id: I31833b4fdcc722730b1bb74d26ac0db235e01717