uboot recv: add delay in polling to allow lower priority tasks to run
The uboot polling task runs in the srtm monitor task, which has a rather
high priority (4, highest after IRQs)

In particular this prevents the CLI from working while waiting for
uboot, and will make things like tty tests fail if we add some later,
so add a short delay between each check like we do for tty rx.

In theory we could also overload the MU0 IRQ handler just for the
duration of the uboot handling and have more efficient processing,
but this is good enough for now.

Change-Id: Ic7a0bfc5d0d2bc50f74d7a94c1896b63ce3fda55