freertos: increase idle task stack size
The idle task is created with configMINIMAL_STACK_SIZE.
If it was just idle, the previous value of 128 would be more than
enough, but it is also doing all the processing for entering and
exiting suspend, and we apparently ran out as can be found out
by resuming from suspend

For the record, the actuall call trace was something like this:
idle
-> vPortSuppressTicksAndSleep
-> APP_PowerPostSwitchHook
-> APP_Resume
-> APP_SRTM_Resume
-> APP_TTY_Resume
-> lpuart_resume (tty_hooks[x]->resume)
-> lpuart_init_device
-> lpuart_setcflag

This can also certainly be improved if we ever need to reclaim that
memory, but for now just increasing the stack trace to 256 is both
good enough and still small enough to not cause memory exhaustion.

Additionally, running a debug build would fail even earlier
(before/during the sleep transition), the exact stack trace has not been
investigated), so this is probably required for debug builds anyway.

Change-Id: Ic7abc53d5786faaa90983fa8e0b38216eccc4de7