GPIO IRQ handler: do not call APP_WakeupACore in irq handler
Running APP_WakeupACore here can cause a race somewhere with upower API,
and we end up with a task blocking waiting for a upower reply it'll never
get.

In theory this might be needed in sleep_mode ignore (or active) modes,
but that turned out not to be required so just remove this code.
If we need it back, we can use APP_SRTM_WakeupCA35 instead which will
schedule a wakeup in the srtm dispatched thread, and will probably avoid
the race

Change-Id: I09bb1f80b954f86dfd936f7ed805c1e664fe6e5a