Add timer to wakeup linux if no MU powerdown irq comes in
If linux aborts suspend after the lfcl suspend notification then we end
up with linux not being able to send notifications to linux anymore.

This should come almost immediately after the notification, so add a
short timer to restore communications the A core does not shutdown.
(This might not work, but at the very least can't be worse than the
current situation)

Change-Id: I4c5e747be9d932f16cf474d004a6ff2a51a54301