IoResume: avoid some glitches with WUU
This avoids WUU firing with MF 0 / PF 0 when it is re-enabled before
sleep on second (deepsleep) suspend onwards.
It looks like WUU remembers that it had a trigger even if the WUU irq is
masked even if we clear PF before re-enabling the irq, disabling it at
PE level avoids this particular issue.

This does not fix the odd behaviour we have in powerdown mode,
neither does keeping WUU enabled all the time; that will require more
investigation.

Change-Id: I8f131ca5c619273a3ca5c61302e92b6aae5bf7f7