suspend: split APP_SRTM_Suspend into EarlySuspend/Suspend
When sleeping with RTOS console opened on linux, it was reported that
the A core voltage would immediately be restored after suspend while
the m33 core (and linux) are still asleep.

It turns out that messages from m33 no longer wake up linux after linux
commit efec0bbad88d ("arm64: dts: armadillo_900: disable wakeup on mu"),
but that was just hiding the problem as the A core would still "wake up"
as far as voltage goes even if linux was not active.

Avoid the problem by splitting suspend in two:
- TTY and CAN (services that can spontaneously send messages to linux)
  are suspeneded early as soon as the suspend lifecycle event comes in
- WDOG (and possibly other services that should be available longer)
  are suspended when m33 core really enters suspend procedure, after A35
  core powered down

If A core didn't power down, LateResume() is called to avoid leaving
armadillo half-suspended.

Change-Id: I8e4b472778ff6f8821036e86e9f046e41209866f