custom/app_tty_custom: improve example tty code
- add comments/split custom_tty_to_linux:
The custom tty code is hard to understand as 'tx' is not tx from the m33
point of view in the custom case; add a few comments and move some code
to a helper that will be reused in demo.
Also check that we only send to linux if the port is open (which is not
a problem for the example code, but will be if custom_tty_to_linux is used
from another task or timer)
- add hook for tty activate
APP might want to know when tty is opened to hook here as well.

Change-Id: Id47de33e36419f577d073fe9d59bf3a8557d1a16