pwm: Add TPM2 and TPM3 support
Enable the use of TPM2 and TPM3, which are peripherals placed in the
Fusion DSP domain.

Following the precedent of TPM0 and TPM1, tried to add the
initialization process for TPM2 and TPM3 in pwm_init_device() of
app_pwm.c. However, since it is executed before the TRDC setting process
after the handshake with U-boot, access to the registers related to TPM2
 and TPM3 was impossible.

Therefore, implement the initialization process for TPM2 and TPM3 in
SRTM, just like SPI and I2C. However, in order to avoid modifying the
PWM source code on the Linux kernel side, introduce it in a rather
abrupt manner by using a flag in the Request of SRTM.

Since the initialization process for TPM0 and TPM1 can be performed
either before or after the TRDC setting, the initialization process is
now performed in the first Request of SRTM together with TPM2 and TPM3.

The handler for the initialization process is written in app_pwm.c.
The interface in adapter, etc. is slightly modified so that the
initialization process is executed in the first Request of each TPM.

Change-Id: I2ad34f05ee7455d6fe2e16812814a34ee1c451f5