pwm: Resolve restrictions on PWM period
The HAL_PwmSetupPwm function and subsequent processing receive the
frequency that is the reciprocal of period as an integer-valued
argument.
A period that corresponds to a non-integer frequency is rounded to an
integer frequency in the process of being converted to an integer
frequency.
Therefore, even if a period corresponding to a non-integer frequency is
specified, it will not actually be that period.

For example, 1000 ms (1 Hz) or 500 ms (2 Hz) will actually be the
specified period, while 800 ms (1.25 Hz) or 0.64 ms (1562.5 Hz) will
actually turn into 1000 ms (1 Hz) or 0.6402... (1562 Hz).

Solve this problem by making a change to accept period directly as an
argument to the HAL_PwmSetupPwm function and subsequent processing.

Change-Id: I11a5245bc53e78d420068cfc4f3f0bbc271c7892