pwm: Improve resolution of duty cycle
In the original mcux-sdk source code, the PWM dutycycle resolution was
limited to 1% of PERIOD (smaller than 7bit).
The i.MX 8ULP has a 32-bit TPM, which does not take full advantage of
its large resolution.

While this is unlikely to be a problem in human-perceived applications
such as LED dimming, 1% resolution is very coarse in situations where
delicate adjustments such as motor control are required.

Therefore, we used ratios expressed from 0 to 2^32 instead.
The reason for setting 2^32 as the maximum instead of 2^32 -1 (32 bits)
is to enable the ratio to be expressed as intended when specifying a
duty cycle that is divisible by a power of 2 for the period.

For example, if the period is set to 200 ns and the duty cycle to
100 ns with a resolution of 25 ns, setting the maximum to 2^32 -1 would
result in the duty ratio not being a neatly divisible value, but
instead being 0.5 smaller than the intended value, effectively setting
the duty cycle to 75 ns.
When 2^32 is set as the maximum, 2^16 is exactly that value, so the
specified 100 ns is set.

Change-Id: I67838af7f3e1ab346712a5a1d7a511bc7ca872e8