pwm: enable/disable PWM with duty cycle
The PWM of each channel of the TPM can be disabled by setting all bits
in the MSB:MSA:ELSB:ELSA fields of the CnSC register for each channel to
0. (As in the original code, setting only the ELSB:ELSA fields to 0 does
not disable the PWM.)
However, with this method, whether the level after disabling becomes
High or Low depends on the timing of the PWM cycle, making it an
operation with uncertainty.

Therefore, following the example of the Linux kernel's pwm-imx-tpm.c
driver, modify the setPwm function so that when it receives a disable
request, it passes a duty cycle of 0 to the subsequent HAL_PwmSetupPwm
function.

Change-Id: I4941266c7a96aec66cbbb2b49077d04e6300c606