pwm: Adjust clock source division for TPM0 and TPM1
With 32-bit PWM (TPM1), the maximum period value ranges from tens to
hundreds of seconds, but since there is little need for such a long
period with hardware PWM, a division factor of 1 was set to maximize the
resolution.

For 16-bit PWM (TPM0), the maximum period value ranges from hundreds of
microseconds to several milliseconds, and the minimum value ranges from
tens of nanoseconds to 100 nanoseconds. However, at this order of
magnitude, longer periods are considered more important than fine
resolution, so the maximum division ratio of 8 was set.

Change-Id: I65be28ddc1b3cc6d2e2b0afb292f363fbf4bf9d1