pwm: Resolving duty cycle errors in EPWM
In Edge-Aligned PWM (EPWM) mode, when setting the period, subtract 1
from the count corresponding to the period and write the result to the
MOD register. Therefore, when calculating the duty cycle value based on
the ratio of duty cycle and MOD, must multiply the ratio of duty cycle
by the value obtained by adding 1 to MOD.

Prior to this commit, the duty cycle ratio was multiplied by MOD,
resulting in a 1-resolution discrepancy between the specified duty cycle
and the actual duty cycle.

Change-Id: Ib3410bac5a5cd84fb480ce02ebfe2b97eed2f87c