pwm: Add support for suspend/resume
TPM2 and TPM3 are instances located on the DSP side, and when the DSP
power is turned off by suspend, TPM2 and TPM3 must be initialized after
resume.

Due to channel setting operations performed before suspend, the actual
initialization process is skipped when initializing with the HAL_PwmInit
function after resume. If initialization is not performed after resume,
a bus fault occurs when operating the SC register, so even if the
HAL_PwmInit function is executed after resume, a bus fault still occurs.

An option was added to force the execution of the internal
initialization process of the HAL_PwmInit function, enabling the use of
TPM2 and TPM3 PWM even after resume.

Change-Id: I818689ad7a13c87b889f0c811340ca2d7014646f